requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
pillow==10.2.0
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
